TestD3D12 : Implemented multi-threaded rendering using the TaskManager

+++ Source/Runtime/Public/Core/Task.h
@@ -34,7 +34,7 @@ public:
     ~TaskManager();
 
                             /// Returns number of threads.
-    size_t                  NumThreads() const { return threads.Count(); }
+    int                     NumThreads() const { return threads.Count(); }
 
                             /// Is task list empty ?
     bool                    IsTaskEmpty() const { return headTaskIndex == tailTaskIndex; }
